+++ lib/praxis/mapper/resource.rb
@@ -4,6 +4,15 @@
 # Once that is complete, the data set is iterated and a resultant view is generated.
 module Praxis
   module Mapper
+    class ResourceNotFound < RuntimeError
+      attr_reader :type, :id
+
+      def initialize(type:, id: nil)
+        @type = type
+        @id = id
+      end
+    end
+
     class Resource
       extend Praxis::Finalizable
 
@@ -58,6 +67,7 @@ module Praxis
         finalize_resource_delegates
         define_model_accessors
 
+        hookup_callbacks
         super
       end
 
@@ -77,6 +87,39 @@ module Praxis
         model._praxis_associations.each do |k, v|
           define_model_association_accessor(k, v) unless instance_methods.include? k
         end
+      end
+
+      def self.hookup_callbacks
+        return unless ancestors.include?(Praxis::Mapper::Resources::Callbacks)
+
+        instance_module = nil
+        class_module = nil
+
+        affected_methods = (before_callbacks.keys + after_callbacks.keys + around_callbacks.keys).uniq
+        affected_methods&.each do |method|
+          calls = {}
+          calls[:before] = before_callbacks[method] if before_callbacks.key?(method)
+          calls[:around] = around_callbacks[method] if around_callbacks.key?(method)
+          calls[:after] = after_callbacks[method] if after_callbacks.key?(method)
+
+          if method.start_with?('self.')
+            # Look for a Class method
+            simple_name = method.to_s.gsub(/^self./, '').to_sym
+            raise "Error building callback: Class-level method #{method} is not defined in class #{name}" unless methods.include?(simple_name)
+
+            class_module ||= Module.new
+            create_override_module(mod: class_module, method: method(simple_name), calls: calls)
+          else
+            # Look for an instance method
+            raise "Error building callback: Instance method #{method} is not defined in class #{name}" unless method_defined?(method)
+
+            instance_module ||= Module.new
+            create_override_module(mod: instance_module, method: instance_method(method), calls: calls)
+          end
+        end
+        # Prepend the created instance and/or class modules if there were any functions in them
+        prepend instance_module if instance_module
+        singleton_class.send(:prepend, class_module) if class_module
       end
 
       def self.for_record(record)
@@ -184,8 +227,11 @@ module Praxis
       end
 
       def self.define_accessor(name)
-        ivar_name = if name.to_s =~ /\?/
+        ivar_name = case name.to_s
+                    when /\?/
                       "is_#{name.to_s[0..-2]}"
+                    when /!/
+                      "#{name.to_s[0..-2]}_bang"
                     else
                       name.to_s
                     end
@@ -250,6 +296,7 @@ module Praxis
       def reload
         clear_memoization
         reload_record
+        self
       end
 
       def clear_memoization
@@ -275,6 +322,81 @@ module Praxis
           super
         end
       end
+
+      # Defines a 'proxy' method in the given module (mod), so it can then be prepended
+      # There are mostly 3 flavors, which dictate how to define the procs (to make sure we play nicely
+      # with ruby's arguments and all). Method with only args, with only kwords, and with both
+      # Note: if procs could be defined with the (...) syntax, this could be more DRY and simple...
+      def self.create_override_module(mod:, method:, calls:)
+        has_args = method.parameters.any? { |(type, _)| %i[req opt rest].include?(type) }
+        has_kwargs = method.parameters.any? { |(type, _)| %i[keyreq keyrest].include?(type) }
+
+        mod.class_eval do
+          if has_args && has_kwargs
+            # Setup the method to take both args and  kwargs
+            define_method(method.name.to_sym) do |*args, **kwargs|
+              calls[:before]&.each do |target|
+                target.is_a?(Symbol) ? send(target, *args, **kwargs) : instance_exec(*args, **kwargs, &target)
+              end
+
+              orig_call = proc { |*a, **kw| super(*a, **kw) }
+              around_chain = calls[:around].inject(orig_call) do |inner, target|
+                proc { |*a, **kw| send(target, *a, **kw, &inner) }
+              end
+              result = if calls[:around].presence
+                         around_chain.call(*args, **kwargs)
+                       else
+                         super(*args, **kwargs)
+                       end
+              calls[:after]&.each do |target|
+                target.is_a?(Symbol) ? send(target, *args, **kwargs) : instance_exec(*args, **kwargs, &target)
+              end
+              result
+            end
+          elsif has_kwargs && !has_args
+            # Setup the method to only take kwargs
+            define_method(method.name.to_sym) do |**kwargs|
+              calls[:before]&.each do |target|
+                target.is_a?(Symbol) ? send(target, **kwargs) : instance_exec(**kwargs, &target)
+              end
+              orig_call = proc { |**kw| super(**kw) }
+              around_chain = calls[:around].inject(orig_call) do |inner, target|
+                proc { |**kw| send(target, **kw, &inner) }
+              end
+              result = if calls[:around].presence
+                         around_chain.call(**kwargs)
+                       else
+                         super(**kwargs)
+                       end
+              calls[:after]&.each do |target|
+                target.is_a?(Symbol) ? send(target, **kwargs) : instance_exec(**kwargs, &target)
+              end
+              result
+            end
+          else
+            # Setup the method to only take args
+            define_method(method.name.to_sym) do |*args|
+              calls[:before]&.each do |target|
+                target.is_a?(Symbol) ? send(target, *args) : instance_exec(*args, &target)
+              end
+              orig_call = proc { |*a| super(*a) }
+              around_chain = calls[:around].inject(orig_call) do |inner, target|
+                proc { |*a| send(target, *a, &inner) }
+              end
+              result = if calls[:around].presence
+                         around_chain.call(*args)
+                       else
+                         super(*args)
+                       end
+              calls[:after]&.each do |target|
+                target.is_a?(Symbol) ? send(target, *args) : instance_exec(*args, &target)
+              end
+              result
+            end
+          end
+        end
+      end
+      private_class_method :create_override_module
     end
   end
 end