+++ spec/praxis/mapper/resource_spec.rb
@@ -32,7 +32,9 @@ describe Praxis::Mapper::Resource do
   context 'retrieving resources' do
     context 'getting a single resource' do
       before do
-        expect(SimpleModel).to receive(:get).with(name: 'george xvi').and_return(record)
+        expect(SimpleModel).to receive(:get) do |args|
+          expect(**args).to match(name: 'george xvi')
+        end.and_return(record)
       end
 
       subject(:resource)  { SimpleResource.get(name: 'george xvi') }
@@ -44,7 +46,9 @@ describe Praxis::Mapper::Resource do
 
     context 'getting multiple resources' do
       before do
-        expect(SimpleModel).to receive(:all).with(name: ['george xvi']).and_return([record])
+        expect(SimpleModel).to receive(:all) do |args|
+          expect(**args).to eq(name: ['george xvi'])
+        end.and_return([record])
       end
 
       subject(:resource_collection) { SimpleResource.all(name: ['george xvi']) }
@@ -174,6 +178,40 @@ describe Praxis::Mapper::Resource do
 
     it 'works regardless of the resource class used' do
       expect(SimpleResource.wrap(record)).to be(OtherResource.wrap(record))
+    end
+  end
+
+  context 'calling typed methods' do
+    let(:resource) { TypedResource }
+    context 'class level ones' do
+      it 'kwarg methods get their args splatted in the top level' do
+        arg = resource.create(name: 'Praxis-licious', payload: { struct_param: { id: 1 } })
+        # Top level args are a hash (cause the typed methods will splat them before calling)
+        expect(arg).to be_kind_of Hash
+        # But structs beyond that are just the loaded types (which we can splat if we want to keep calling)
+        expect(arg[:payload]).to be_kind_of Attributor::Struct
+      end
+
+      it 'non-kwarg methods get a single arg' do
+        arg = resource.singlearg_create({ name: 'Praxis-licious', payload: { struct_param: { id: 1 } } })
+        # Single argument, instance of an Attributor Struct
+        expect(arg).to be_kind_of Attributor::Struct
+      end
+    end
+    context 'instance level ones' do
+      it 'kwarg methods get their args splatted in the top level' do
+        arg = resource.new({}).update!(string_param: 'Stringer', struct_param: { id: 1 })
+        # Top level args are a hash (cause the typed methods will splat them before calling)
+        expect(arg).to be_kind_of Hash
+        # But structs beyond that are just the loaded types (which we can splat if we want to keep calling)
+        expect(arg[:struct_param]).to be_kind_of Attributor::Struct
+      end
+
+      it 'non-kwarg methods get a single arg' do
+        arg = resource.new({}).singlearg_update!({ string_param: 'Stringer', struct_param: { id: 1 } })
+        # Single argument, instance of an Attributor Struct
+        expect(arg).to be_kind_of Attributor::Struct
+      end
     end
   end
 end